refactor(css): split css module parser generator [skip ci]
